+++ src/components/mod.rs
@@ -1,9 +1,11 @@
+mod b_breadcrumb;
 mod b_dropdown;
 mod b_menu;
 mod b_modal;
 mod b_navbar;
 mod b_pagination;
 
+pub use b_breadcrumb::*;
 pub use b_dropdown::*;
 pub use b_menu::*;
 pub use b_modal::*;