+++ website/src/layout.rs
@@ -88,7 +88,9 @@ pub fn Layout(children: Children) -> impl IntoView {
                         <BNavbarItem href="/elements#button">{t!(i18n, button)}</BNavbarItem>
                         <BNavbarItem href="/elements/tag">{t!(i18n, tag)}</BNavbarItem>
                     </BNavbarItemDropdown>
-                    <BNavbarItem href="/components">{t!(i18n, components)}</BNavbarItem>
+                    <BNavbarItemDropdown is_hoverable=true href="/components" trigger=move || t!(i18n, components)>
+                        <BNavbarItem href="/components/breadcrumb">{t!(i18n, breadcrumb)}</BNavbarItem>
+                    </BNavbarItemDropdown>
                     <BNavbarItem href="/form">{t!(i18n, form)}</BNavbarItem>
                     <BNavbarItem href="/columns">{t!(i18n, columns)}</BNavbarItem>
                     <BNavbarItem href="/layout">{t!(i18n, layout)}</BNavbarItem>