+++ website/src/pages/components/mod.rs
@@ -1,0 +1,3 @@
+mod breadcrumb_page;
+
+pub use breadcrumb_page::BreadcrumbPage;